requirements.txt
futures==2.2.0
kafka-python==2.0.2
requests==2.31.0
urllib3==2.2.1